requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
chromadb==0.5.3
rank-bm25==0.2.2
openai==1.35.3
streamlit==1.36.0
pandas==2.2.2
openpyxl==3.1.4
plotly==5.22.0
